requirements/production.txt
-r ./base.txt

gunicorn==19.9.0

# Static files
django-storages[boto3]==1.7.1

# Email
django-anymail[mailgun]==5.0